wyze_bridge/requirements.txt
fastapi
uvicorn[standard]
asyncio
wyze-sdk
python-dotenv